feat: Modify combinatorial reasoning component from Icosa
