+++ src/backend/base/langflow/components/icosacomputing/combinatorial_reasoner.py
@@ -1,50 +1,79 @@
 import requests
-from requests.auth import HTTPBasicAuth
 
 from langflow.base.models.openai_constants import OPENAI_MODEL_NAMES
 from langflow.custom import Component
-from langflow.inputs import DropdownInput, SecretStrInput, StrInput
+from langflow.field_typing.range_spec import RangeSpec
+from langflow.inputs import DropdownInput, SecretStrInput, SliderInput, StrInput
 from langflow.io import MessageTextInput, Output
-from langflow.schema import Data
 from langflow.schema.message import Message
+
+ICOSA_OPTIONS = ["Poll", "Reason (Beta)"]
 
 
 class CombinatorialReasonerComponent(Component):
     display_name = "Combinatorial Reasoner"
-    description = "Uses Combinatorial Optimization to construct an optimal prompt with embedded reasons. Sign up here:\nhttps://forms.gle/oWNv2NKjBNaqqvCx6"
+    description = (
+        "Optimize the “thinking” of models at test-time to increase accuracy. Uses Icosa models by default "
+        "but can run on any model by changing base URL in Controls. Sign up for free at icosacomputing.com"
+    )
     icon = "Icosa"
     name = "Combinatorial Reasoner"
 
     inputs = [
         MessageTextInput(name="prompt", display_name="Prompt", required=True),
         SecretStrInput(
-            name="openai_api_key",
-            display_name="OpenAI API Key",
-            info="The OpenAI API Key to use for the OpenAI model.",
+            name="icosa_api_key",
+            display_name="Icosa API Key",
+            info="Sign up at icosacomputing.com for this key",
             advanced=False,
-            value="OPENAI_API_KEY",
+            value="ICOSA_API_KEY",
             required=True,
         ),
         StrInput(
-            name="username",
-            display_name="Username",
-            info="Username to authenticate access to Icosa CR API",
-            advanced=False,
-            required=True,
-        ),
-        SecretStrInput(
-            name="password",
-            display_name="Password",
-            info="Password to authenticate access to Icosa CR API.",
-            advanced=False,
-            required=True,
-        ),
-        DropdownInput(
             name="model_name",
             display_name="Model Name",
-            advanced=False,
-            options=OPENAI_MODEL_NAMES,
+            advanced=True,
             value=OPENAI_MODEL_NAMES[0],
+        ),
+        DropdownInput(
+            name="prompt_type",
+            display_name="Type",
+            advanced=False,
+            info=(
+                "(Poll: After sampling N times, gives the most likely answer in Optimized Response. "
+                "Statistically relevant reasons are given in Selected Reasons. Reason (Beta): Instead of "
+                "polling, directly uses Selected Reasons to guide the model and get the answer, given in "
+                "Optimized Response."
+            ),
+            options=ICOSA_OPTIONS,
+            value=ICOSA_OPTIONS[0],
+        ),
+        SliderInput(
+            name="num_samples",
+            display_name="Compute Power (Samples)",
+            info=(
+                "Number of samples to use for the optimization. More samples implies more accurate results, "
+                "but increased compute."
+            ),
+            min_label="",
+            max_label="",
+            value=15,
+            range_spec=RangeSpec(min=1, max=30, step=1, step_type="int"),
+        ),
+        SecretStrInput(
+            name="openai_api_key",
+            display_name="OpenAI API Key",
+            info="The API Key to use for the OpenAI model.",
+            advanced=True,
+            value="OPENAI_API_KEY",
+            required=False,
+        ),
+        StrInput(
+            name="base_url",
+            display_name="Base URL",
+            info="Call any provider compatible with OpenAI's SDK by providing a base URL",
+            advanced=True,
+            value="",
         ),
     ]
 
@@ -54,6 +83,11 @@ class CombinatorialReasonerComponent(Component):
             name="optimized_prompt",
             method="build_prompt",
         ),
+        Output(
+            display_name="Response",
+            name="response",
+            method="build_response",
+        ),
         Output(display_name="Selected Reasons", name="reasons", method="build_reasons"),
     ]
 
@@ -62,23 +96,27 @@ class CombinatorialReasonerComponent(Component):
             "prompt": self.prompt,
             "apiKey": self.openai_api_key,
             "model": self.model_name,
+            "type": self.prompt_type,
+            "numSamples": self.num_samples,
+            "baseUrl": self.base_url,
         }
-
-        creds = HTTPBasicAuth(self.username, password=self.password)
+        headers = {"X-API-Key": self.icosa_api_key}
         response = requests.post(
-            "https://cr-api.icosacomputing.com/cr/langflow",
+            "https://cr-api.icosacomputing.com/cr/langflow_lite",
             json=params,
-            auth=creds,
+            headers=headers,
             timeout=100,
         )
         response.raise_for_status()
 
         prompt = response.json()["prompt"]
-
         self.reasons = response.json()["finalReasons"]
+        self.response = response.json()["answerWithReasoning"]
         return prompt
 
-    def build_reasons(self) -> Data:
-        # list of selected reasons
+    def build_response(self) -> Message:
+        return self.response
+
+    def build_reasons(self) -> Message:
         final_reasons = [reason[0] for reason in self.reasons]
-        return Data(value=final_reasons)
+        return Message(text="\n".join(final_reasons))
