+++ src/backend/base/langflow/initial_setup/starter_projects/Financial Agent.json
@@ -2257,7 +2257,9 @@
                 "display_name": "Message",
                 "method": "text_response",
                 "name": "text_output",
-                "required_inputs": [],
+                "required_inputs": [
+                  "api_key"
+                ],
                 "selected": "Message",
                 "tool_mode": true,
                 "types": [
