+++ src/backend/base/langflow/initial_setup/starter_projects/Financial Report Parser.json
@@ -160,7 +160,9 @@
                 "display_name": "Message",
                 "method": "text_response",
                 "name": "text_output",
-                "required_inputs": [],
+                "required_inputs": [
+                  "api_key"
+                ],
                 "selected": "Message",
                 "tool_mode": true,
                 "types": [
