+++ src/backend/base/langflow/initial_setup/starter_projects/Market Research.json
@@ -2197,7 +2197,9 @@
                 "display_name": "Message",
                 "method": "text_response",
                 "name": "text_output",
-                "required_inputs": [],
+                "required_inputs": [
+                  "api_key"
+                ],
                 "selected": "Message",
                 "tool_mode": true,
                 "types": [
