+++ src/backend/base/langflow/initial_setup/starter_projects/Vector Store RAG.json
@@ -2810,7 +2810,9 @@
                 "display_name": "Message",
                 "method": "text_response",
                 "name": "text_output",
-                "required_inputs": [],
+                "required_inputs": [
+                  "api_key"
+                ],
                 "selected": "Message",
                 "tool_mode": true,
                 "types": [
@@ -3218,7 +3220,11 @@
                 "display_name": "DataFrame",
                 "method": "as_dataframe",
                 "name": "dataframe",
-                "required_inputs": [],
+                "required_inputs": [
+                  "collection_name",
+                  "database_name",
+                  "token"
+                ],
                 "selected": "DataFrame",
                 "tool_mode": true,
                 "types": [
@@ -3949,7 +3955,11 @@
                 "display_name": "DataFrame",
                 "method": "as_dataframe",
                 "name": "dataframe",
-                "required_inputs": [],
+                "required_inputs": [
+                  "collection_name",
+                  "database_name",
+                  "token"
+                ],
                 "selected": "DataFrame",
                 "tool_mode": true,
                 "types": [
